20191006 2324 Adding Ranking Data to Scheduling Data
Beginning to add the ranking data into the scheduling data.  Currently
the data is merged in with some minor bugs, like not being able to
show the preseason ranking for games in week 1.  This is probably
also an issue for games in the postseason.  Also note that the ranking
data is not showing up in the emails being sent to the To: address
that is listing in the config.ini file.